third_party/WebKit/LayoutTests/platform/win/external/wpt/referrer-policy/generic/subresource-test/fetch-messaging-expected.txt

This is a testharness.js-based test.
PASS Fetch is supported by the browser. 
PASS Fetch is responding with HTTP headers 
Harness: the test ran to completion.
